+++ AtmosphereAutopilot/AtmosphereAutopilot.cs
@@ -22,7 +22,6 @@ using System.Text;
 using UnityEngine;
 using System.Reflection;
 using KSP.UI.Screens;
-using AtmosphereAutopilot.UI;
 //using ToolbarWrapper;
 
 namespace AtmosphereAutopilot
@@ -276,7 +275,7 @@ namespace AtmosphereAutopilot
         private Texture launcher_btn_textore_off = null;
         private Texture launcher_btn_textore_on = null;
 
-        private UI.MainMenu toolbar_menu = null;
+        private UI.MainMenuGUI toolbar_menu = null;
         private GameObject toolbar_menu_object = null;
         private GameObject toolbar_menu_prefab = null;
 
@@ -332,15 +331,6 @@ namespace AtmosphereAutopilot
                 mainMenuClose();
         }
 
-        private void setLauncherOnOffIcon (bool state) {
-            if (launcher_btn == null)
-                return;
-            if (state)
-                launcher_btn.SetTexture (launcher_btn_textore_on);
-            else 
-                launcher_btn.SetTexture (launcher_btn_textore_off);
-        }
-
         private void mainMenuClose() {
             if (toolbar_menu != null)
                 toolbar_menu.fadeOut();
@@ -366,9 +356,9 @@ namespace AtmosphereAutopilot
             return;
 
           toolbar_menu_object.transform.SetParent(MainCanvasUtil.MainCanvas.transform);
-          toolbar_menu = toolbar_menu_object.GetComponent<UI.MainMenu> ();
+          toolbar_menu = toolbar_menu_object.GetComponent<UI.MainMenuGUI> ();
           if (toolbar_menu != null)
-            toolbar_menu.setController(new MainMenuController (this));
+            toolbar_menu.setController(new NeoGUIController (this));
           GUIStyles.Process(toolbar_menu_object);
         }
 
@@ -383,6 +373,21 @@ namespace AtmosphereAutopilot
             return anchor;
         }
 
+        public void setLauncherOnOffIcon (bool state) {
+            if (launcher_btn == null)
+                return;
+            if (state)
+                launcher_btn.SetTexture (launcher_btn_textore_on);
+            else 
+                launcher_btn.SetTexture (launcher_btn_textore_off);
+        }
+
+        public bool launcherButtonState {
+            get {
+                return launcher_btn != null && launcher_btn.toggleButton.CurrentState == KSP.UI.UIRadioButton.State.True;
+            }
+        }
+
         public void mainMenuGUIUpdate () {
             if (toolbar_menu != null)
               toolbar_menu.updateGUI ();
@@ -390,256 +395,6 @@ namespace AtmosphereAutopilot
                 autopilot_module_lists[ActiveVessel].ContainsKey(typeof(TopModuleManager)))
               setLauncherOnOffIcon (autopilot_module_lists[ActiveVessel][typeof(TopModuleManager)].Active);
         }
-
-    private class MainMenuController : UI.IGUIController {
-      private AtmosphereAutopilot parent;
-
-      private TopModuleManager masterAP = null;
-      private StandardFlyByWire fbwAP = null;
-      private CruiseController ccAP = null;
-      private MouseDirector dcAP = null;
-      private ProgradeThrustController speedAP = null;
-      private PitchAngularVelocityController pvc = null;
-      private YawAngularVelocityController yvc = null;
-
-      private Autopilots currentAP {
-        get {
-          findModules ();
-          if (masterAP != null && masterAP.Active == false)
-            return Autopilots.DISABLED;
-          if (fbwAP != null && fbwAP.Active)
-            return Autopilots.FLYBYWIRE;
-          if (ccAP != null && ccAP.Active)
-            return Autopilots.CRUISECTRL;
-          if (dcAP != null && dcAP.Active)
-            return Autopilots.MOUSEDIR;
-          return Autopilots.DISABLED;
-        }
-      }
-
-      private void findModules () {
-        if (parent.ActiveVessel == null || parent.autopilot_module_lists.ContainsKey (parent.ActiveVessel) == false)
-          return;
-
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (TopModuleManager)))
-          masterAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (TopModuleManager)] as TopModuleManager;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (StandardFlyByWire)))
-          fbwAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (StandardFlyByWire)] as StandardFlyByWire;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (CruiseController)))
-          ccAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (CruiseController)] as CruiseController;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (MouseDirector)))
-          dcAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (MouseDirector)] as MouseDirector;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (ProgradeThrustController)))
-          speedAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (ProgradeThrustController)] as ProgradeThrustController;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (PitchAngularVelocityController)))
-          pvc = parent.autopilot_module_lists[parent.ActiveVessel][typeof (PitchAngularVelocityController)] as PitchAngularVelocityController;
-        if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (YawAngularVelocityController)))
-          yvc = parent.autopilot_module_lists[parent.ActiveVessel][typeof (YawAngularVelocityController)] as YawAngularVelocityController;
-      }
-
-      internal MainMenuController (AtmosphereAutopilot parent) {
-        this.parent = parent;
-
-        findModules ();
-      }
-
-      public bool appButtonOn {
-        get {
-          return parent.launcher_btn != null && parent.launcher_btn.toggleButton.CurrentState == KSP.UI.UIRadioButton.State.True;
-        }
-      }
-
-      public Autopilots currentAutopilot {
-        get {
-          return currentAP;
-        }
-
-        set {
-          if (masterAP == null)
-            return;
-
-          switch (value) {
-            case Autopilots.DISABLED:
-              fbwAP = null;
-              ccAP = null;
-              dcAP = null;
-              speedAP = null;
-              pvc = null;
-              yvc = null;
-              masterAP.Active = false;
-              parent.setLauncherOnOffIcon (false);
-              return;
-            case Autopilots.FLYBYWIRE:
-              fbwAP = masterAP.activateAutopilot (typeof (StandardFlyByWire)) as StandardFlyByWire;
-              ccAP = null;
-              dcAP = null;
-              break;
-            case Autopilots.CRUISECTRL:
-              fbwAP = null;
-              ccAP = masterAP.activateAutopilot (typeof (CruiseController)) as CruiseController;
-              dcAP = null;
-              break;
-            case Autopilots.MOUSEDIR:
-              fbwAP = null;
-              ccAP = null;
-              dcAP = masterAP.activateAutopilot (typeof (MouseDirector)) as MouseDirector;
-              break;
-          }
-          if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (ProgradeThrustController)))
-            speedAP = parent.autopilot_module_lists[parent.ActiveVessel][typeof (ProgradeThrustController)] as ProgradeThrustController;
-          if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (PitchAngularVelocityController)))
-            pvc = parent.autopilot_module_lists[parent.ActiveVessel][typeof (PitchAngularVelocityController)] as PitchAngularVelocityController;
-          if (parent.autopilot_module_lists[parent.ActiveVessel].ContainsKey (typeof (YawAngularVelocityController)))
-            yvc = parent.autopilot_module_lists[parent.ActiveVessel][typeof (YawAngularVelocityController)] as YawAngularVelocityController;
-
-          parent.setLauncherOnOffIcon (masterAP.Active);
-        }
-      }
-
-      public bool speedControl {
-        get {
-          return speedAP != null && speedAP.spd_control_enaled;
-        }
-
-        set {
-          if (speedAP != null)
-            speedAP.spd_control_enaled = value;
-        }
-      }
-
-      public float speed {
-        get {
-          if (speedAP == null)
-            return 0f;
-          return speedAP.spd_setpoint;
-        }
-
-        set {
-          if (parent.ActiveVessel != null)
-            speedAP.spd_setpoint = value;
-        }
-      }
-
-      // fly-by-wire
-      public bool rocketMode {
-        get {
-          return fbwAP != null && fbwAP.rocket_mode;
-        }
-        set {
-          if (fbwAP != null)
-            fbwAP.rocket_mode = value;
-        }
-      }
-
-      // the "normal" value for all the limits, hardcoded
-      private const float moderationNorm = 15f;
-
-      private void setModerationLimits (float limit) {
-        if (pvc != null) {
-          pvc.max_aoa = limit;
-          pvc.max_g_force = limit;
-        }
-        if (yvc != null) {
-          yvc.max_aoa = limit;
-          yvc.max_g_force = limit;
-        }
-      }
-
-      public bool moderation {
-        get {
-          return fbwAP != null && fbwAP.moderation_switch;
-        }
-        set {
-          if (fbwAP != null)
-            fbwAP.moderation_switch = value;
-        }
-      }
-
-      public float moderationMult {
-        get {
-          if (pvc == null)
-            return 1f;
-          //take pitch aoa limit as the master
-          float value = pvc.max_aoa;
-          setModerationLimits (value);
-          return value / moderationNorm;
-        }
-        set {
-          setModerationLimits (value * moderationNorm);
-        }
-      }
-
-      // cruise control
-      public bool altitudeControl {
-        get {
-          return ccAP != null && ccAP.vertical_control;
-        }
-        set {
-          if (ccAP != null)
-            ccAP.vertical_control = value;
-        }
-      }
-
-      public float altitude {
-        get {
-          if (ccAP == null)
-            return 0f;
-          return ccAP.desired_altitude;
-        }
-        set {
-          if (ccAP != null)
-            ccAP.desired_altitude = value;
-        }
-      }
-
-      public float distToWaypoint {
-        get {
-          if (ccAP == null || ccAP.current_mode != CruiseController.CruiseMode.Waypoint)
-            return -1f;
-
-          double straightDist = ccAP.dist_to_dest;
-          // no corrections if the target is near
-          if (straightDist <= 10000.0)
-            return (float)straightDist;
-          // calculate the surface distance
-          double radius = parent.ActiveVessel.mainBody.Radius;
-          return (float)(Math.Acos (1 - (straightDist * straightDist) / (2 * radius * radius)) * radius);
-        }
-      }
-
-      public bool pickingWaypoint {
-        get {
-          return ccAP != null && ccAP.current_mode == CruiseController.CruiseMode.Waypoint && ccAP.picking_waypoint;
-        }
-      }
-
-      public bool waypointIsSet {
-        get {
-          return ccAP != null && ccAP.current_mode == CruiseController.CruiseMode.Waypoint && ccAP.waypoint_entered;
-        }
-      }
-
-      public void pickWaypoint () {
-        if (ccAP == null)
-          return;
-        ccAP.WaypointMode = true;
-        MapView.EnterMapView ();
-        ccAP.picking_waypoint = true;
-      }
-
-      public void cancelWaypointPick () {
-        if (ccAP == null)
-          return;
-        ccAP.picking_waypoint = false;
-      }
-
-      public void removeWaypoint () {
-        if (ccAP == null)
-          return;
-        ccAP.picking_waypoint = false;
-        ccAP.LevelFlightMode = true;
-      }
-    }
 
         #endregion
