+++ AtmosphereAutopilot/AtmosphereAutopilot.csproj
@@ -71,6 +71,7 @@
     <None Include="ToolbarWrapper.cs" />
     <None Include="Controller\PDController.cs" />
     <Compile Include="GimbalRearranger.cs" />
+    <Compile Include="GUI\NeoGUIController.cs" />
     <Compile Include="ListSelector.cs" />
     <Compile Include="ListView.cs" />
     <None Include="Math\GradientDescend.cs" />
