+++ AtmosphereAutopilot/Modules/CruiseController.cs
@@ -94,6 +94,9 @@ namespace AtmosphereAutopilot
             if (vessel.LandedOrSplashed)
                 return;
 
+            if (thrust_c.spd_control_enaled)  
+                thrust_c.ApplyControl(cntrl, thrust_c.spd_setpoint);  
+
             desired_velocity = Vector3d.zero;
             planet2ves = vessel.ReferenceTransform.position - vessel.mainBody.position;
             planet2vesNorm = planet2ves.normalized;
