+++ AtmosphereAutopilot/NeoGUI/Utils/CanvasGroupFader.cs
