+++ AtmosphereAutopilot/NeoGUI/Utils/ShrinkPanel.cs
