+++ AtmosphereAutopilot/NeoGUI/Utils/StyleApplicator.cs
